// src/api/openapi/models/virtual-account-info.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Zeroone API
 * Class APIs are served under /api/v5 while legacy domains may still remain on /api/v1 during migration.
 *
 * The version of the OpenAPI document: v5-class-migration
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface VirtualAccountInfo {
    'accountNumber'?: string;
    'bankCode'?: string;
    'bankName'?: string;
    'customerName'?: string;
    'dueDate'?: string;
    'expired'?: boolean;
}